+++ phpunit/media/media-processing-test.php
@@ -123,6 +123,7 @@ class Media_Processing_Test extends WP_UnitTestCase {
 		$this->assertArrayNotHasKey( 'image_sizes', $data );
 		$this->assertArrayNotHasKey( 'image_strip_meta', $data );
 		$this->assertArrayNotHasKey( 'image_max_bit_depth', $data );
+		$this->assertArrayNotHasKey( 'video_keep_original', $data );
 	}
 
 	/**
@@ -158,6 +159,8 @@ class Media_Processing_Test extends WP_UnitTestCase {
 		$this->assertTrue( $data['image_strip_meta'] );
 		$this->assertArrayHasKey( 'image_max_bit_depth', $data );
 		$this->assertSame( 16, $data['image_max_bit_depth'] );
+		$this->assertArrayHasKey( 'video_keep_original', $data );
+		$this->assertTrue( $data['video_keep_original'] );
 	}
 
 	/**